fix: The Paint component now works even if you scale it through CSS.
This change was done using ChatGPT :-D ... I just explained the issue and uploaded the file. This was faster than analyzing the code, but might have lead to silliness. I think the solution is probably fine and it's nothing too crazy, but I haven't done a thorough examination of the issue so sorry if it's silly.